Module Impl.Inspect_durable_state

Inspect durable state using a more specialised way of reading the PVM state. For example in WASM, it decodes the durable storage in the state before reading values.

val lookup : state -> string list -> bytes option Lwt.t

lookup state path returns the data stored for the path path in the PVM state state.