Module Octez_smart_rollup_node_PtNairob.Inbox

The rollup node maintains an inbox of incoming messages.

The incoming messages for a rollup are published on the layer 1. To maintain the state of its inbox, a rollup node retrieves these messages each time the tezos blockchain is updated.

The inbox state is persistent.

val process_head : Octez_smart_rollup_node.Node_context.rw -> predecessor:Octez_smart_rollup_node.Layer1.header -> Octez_smart_rollup_node.Layer1.header -> (Octez_smart_rollup.Inbox.Hash.t * Octez_smart_rollup.Inbox.t * Octez_smart_rollup.Merkelized_payload_hashes_hash.t * string list) Tezos_base.TzPervasives.tzresult Lwt.t

process_head node_ctxt ~predecessor head operations changes the state of the inbox to react to head (where predecessor is the predecessor of head in the L1 chain). In particular, this process filters the provided operations of the head block.

val add_messages : is_first_block:bool -> predecessor_timestamp: Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Timestamp.time -> predecessor:Tezos_base.TzPervasives.Block_hash.t -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Inbox.t -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Inbox_message.t list -> (Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Inbox_merkelized_payload_hashes.History.t * Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Inbox_merkelized_payload_hashes.Hash.t * Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Inbox.t * Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Inbox_message.t list) Tezos_base.TzPervasives.tzresult Lwt.t

add_messages ~is_first_block ~predecessor_timestamp ~predecessor inbox messages adds messages to the inbox using Sc_rollup.Inbox.add_all_messages.

val payloads_history_of_messages : is_first_block:bool -> predecessor:Tezos_base.TzPervasives.Block_hash.t -> predecessor_timestamp: Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Timestamp.time -> string list -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Inbox_merkelized_payload_hashes.History.t Tezos_base.TzPervasives.tzresult

payloads_history_of_messages ~is_first_block ~predecessor ~predecessor_timestamp messages builds the payloads history for the list of messages. This allows to not store payloads histories (which contain merkelized skip lists) but simply messages.

val payloads_history_of_all_messages : string list -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Inbox_merkelized_payload_hashes.History.t Tezos_base.TzPervasives.tzresult

payloads_history_of_all_messages messages builds the merkelized payloads history for the list of serialzied messages messages.

val same_as_layer_1 : _ Octez_smart_rollup_node.Node_context.t -> Tezos_base.TzPervasives.Block_hash.t -> Octez_smart_rollup.Inbox.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

same_as_layer_1 node_ctxt block node_inbox ensures that the rollup node agrees with the L1 node that inbox for block is node_inbox.

val serialize_external_message : string -> string Tezos_base.TzPervasives.tzresult

Serialize an external messages to the protocol representation. NOTE: this adds a tag '\001' at the beginning.

val init : predecessor_timestamp:Tezos_base.TzPervasives.Time.Protocol.t -> predecessor:Tezos_base.TzPervasives.Block_hash.t -> level:int32 -> Octez_smart_rollup.Inbox.t

Returns the initial global inbox where level is the first level of the protocol with smart rollups.