Module Octez_smart_rollup_node_PtNairob.Layer1_event

This module defines functions that emit the events used by the layer 1 chain (see Layer).

val starting : unit -> unit Lwt.t

val stopping : unit -> unit Lwt.t

val switched_new_head : Tezos_base.TzPervasives.Block_hash.t -> int32 -> unit Lwt.t

switched_new_head hash level emits the event that the layer 1 has notified a new head with hash at some given level.