Module S.Unsafe_patches

Expose unsafe state patching functions for manual intervention. At the moment this feature is only used to increase the maximum number of ticks of the WASM PVM in a non refutable setting.

type t

val of_patch : Octez_smart_rollup_node.Pvm_patches.unsafe_patch -> t Tezos_base.TzPervasives.tzresult

of_patch p returns the PVM patch if it has a corresponding one.

val apply : state -> t -> state Lwt.t

apply state patch applies the unsafe patch patch on the state.